requirements.txt
dvc[gdrive]==3.60.1
hydra-core==1.3.2
lightning==2.5.1
mlflow==3.1.1
pandas==2.3.0
scikit-image==0.25.2
torch==2.7.1
torchvision==0.22.1
